+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Develop
 
+- Rework library CMake with removed INTERFACE type
+
 ## 1.2.0
 
 - Change license year to 2022